+++ projects/ui/src/public-api.ts
@@ -76,4 +76,5 @@ export * from './lib/combobox/combobox';
 export * from './lib/layout/layout';
 export * from './lib/bento/bento';
 export * from './lib/surface/surface';
+export * from './lib/calendar/calendar';
 export * from './lib/version';